chore: rename trends module
