+++ stitcher.py
